RobustVerification.jl

Walk through: a first iteration

Caveats

Graph prototyping version 01

Graph prototyping version 02

Test redo

graph construction and coloring

Graph prototyping version 02

RobustVerification.jl

Structs

Graph based functions

Plot graph colourings

Round functions

Redo function

Compute angle function

Helper functions

Index

Learning to use BraKets in Julia

Protocol prototype version 01

Synbolics

Structs

RobustVerificationSimulator.Redo — Type

Redo round type

RobustVerificationSimulator.Test — Type

Round Test type

RobustVerificationSimulator.Computation — Type

Round Computation type

RobustVerificationSimulator.Trap — Type

Qubit Trap type

RobustVerificationSimulator.Dummy — Type

Qubit Dummy type

RobustVerificationSimulator.ComputationQubit — Type

Qubit ComputationQubit type

RobustVerificationSimulator.Input — Type

Qubit Input type

RobustVerificationSimulator.NotInput — Type

Qubit NotInput type

RobustVerificationSimulator.Ok — Type

Round outcome type

RobustVerificationSimulator.Abort — Type

Round Abort type

Graph based functions

RobustVerificationSimulator.generate_simple_graph — Function

From number of vertices and edges (v,e)-> Graph

RobustVerificationSimulator.generate_random_greedy_color — Function

Use Graphs.randomgreedycolor(g, reps) -> Graphs.Coloring{Int64} Uses a greedy approximation to get a colouring of a graph

RobustVerificationSimulator.separate_each_color — Function

Extracts from a Graphs.Coloring{Int64} -> Vector{Vector{Int64}} Once a coloring is selected a vector of integers will result 1 -> Dummy vertex 2 -> Trap

RobustVerificationSimulator.get_random_coloring — Function

Extracts from Vector{Vector{Int64}} -> Vector{Int64} Draw uniform one colouring arangments

RobustVerificationSimulator.generate_property_graph! — Function

From (MetaGraph{Int64, 
    Union{Test,Computation},
    Float64},Vector{Int64}) - >
        MetaGraph{Int64, Float64}
The random approximate independnet set  is used to label the graph
According to the Trap and Dummy
This function affects the input meta graph so it had a !
For Trap a random angle is sent over
For Dummy random 0 or 1

From (MetaGraph{Int64, Float64},Union{Test,Computation}) - >
        MetaGraph{Int64, Float64}
Take SimpleGraph, just converted to MetaGraph and add angles.
Note: this may be a temporary solutions, I need to implement 
an input and output set for the vertices

Plot graph colourings

RobustVerificationSimulator.contruct_coloring_plot_for_one_color — Function

Plot graph g with the random coloring stipulated from functions:
1. generate_random_greedy_color
2. g_col = get_random_coloring

RobustVerificationSimulator.contruct_coloring_plot_for_all_colors — Function

Plot graph g with the random coloring stipulated from functions:
1. generate_random_greedy_color
2. g_col = get_random_coloring

RobustVerificationSimulator.contruct_coloring_plot_for_no_colors — Function

Plot graph whith no coloring

Round functions

Including functions dealing with Computation or Test and further for Trap or Dummy qubit tyes.

RobustVerificationSimulator.choose_test_computation — Function

Input a probability -> Union{Test,Computation}

RobustVerificationSimulator.return_ok_or_abort — Function

Input a probability -> Union{OK,Abort}

RobustVerificationSimulator.get_round_result — Function

Take (round_type,p) -> Union{OK,Abort}

RobustVerificationSimulator.qubit_information — Function

Trap() -> get kπ/4 k∈0..7

Dummy() -> 0,1

Computation() -> get kπ/4 k∈0..7

RobustVerificationSimulator.compare_trap_test — Function

Function, which will iterate over vertices
will yield if the round passed or failed

Redo function

RobustVerificationSimulator.do_redo — Function

Input two proabailites (u,p) -> u
If uₜₑₛₜ < p -> draw new u ~ U(0,1)
else return u

Compute angle function

RobustVerificationSimulator.compute_δᵥ — Function

Computation of δᵥ Case

Round ≡ Test ∩ Qubit ≡ Dummy → δᵥ = {kπ/r | k ∼ U(0..7)}

Computation of δᵥ Case

Round ≡ Test ∩ Qubit ≡ Trap → δᵥ = θᵥ + rᵥπ

Computation of δᵥ Case

Round ≡ Computation ∩ Qubit ∈ Input set → δᵥ = ϕᵥ + (θᵥ + xᵥπ) + rᵥπ

Computation of δᵥ Case

Round ≡ Computation ∩ Qubiut ∉ Input set → δᵥ = ϕᵥ′ + θᵥ + rᵥπ

Temporary function till I implement/figure out the input/output set Computation of δᵥ Case

Round ≡ Computation ∩ Qubiut ∉ Input set → δᵥ = ϕᵥ′ + θᵥ + rᵥπ

Helper functions

RobustVerificationSimulator.get_ratio_between_two_types — Function

Get ratio of occurence between two types

RobustVerificationSimulator.rand_int — Function

Draw random integer either 0 or 1

RobustVerificationSimulator.probability_asser — Function

Assert that a number is between 0 and 1 and asserts that 
this true.

RobustVerificationSimulator.error_call_not_test_computation — Function

Call if the condition of Test or Computation is not meant

Index

RobustVerificationSimulator.Abort

RobustVerificationSimulator.Computation

RobustVerificationSimulator.ComputationQubit

RobustVerificationSimulator.Dummy

RobustVerificationSimulator.Input

RobustVerificationSimulator.NotInput

RobustVerificationSimulator.Ok

RobustVerificationSimulator.Redo

RobustVerificationSimulator.Test

RobustVerificationSimulator.Trap

RobustVerificationSimulator.choose_test_computation

RobustVerificationSimulator.compare_trap_test

RobustVerificationSimulator.compute_δᵥ

RobustVerificationSimulator.contruct_coloring_plot_for_all_colors

RobustVerificationSimulator.contruct_coloring_plot_for_no_colors

RobustVerificationSimulator.contruct_coloring_plot_for_one_color

RobustVerificationSimulator.do_redo

RobustVerificationSimulator.error_call_not_test_computation

RobustVerificationSimulator.generate_property_graph!

RobustVerificationSimulator.generate_random_greedy_color

RobustVerificationSimulator.generate_simple_graph

RobustVerificationSimulator.get_random_coloring

RobustVerificationSimulator.get_ratio_between_two_types

RobustVerificationSimulator.get_round_result

RobustVerificationSimulator.probability_asser

RobustVerificationSimulator.qubit_information

RobustVerificationSimulator.rand_int

RobustVerificationSimulator.return_ok_or_abort

RobustVerificationSimulator.separate_each_color